FSAL_VFS: fix clang compile warning/error os/linux/handle_syscalls.c
(compiling with clang on Linux)

default build options -Wall -Werror -Wformat results in

.../FSAL_VFS/os/linux/handle_syscalls.c:429:5: error: format specifies type 'unsigned char' but the argument has type 'int' [-Werror,-Wformat]
                         handle0 & HANDLE_FSID_MASK);
                         ^~~~~~~~~~~~~~~~~~~~~~~~~~

also at line 460.

+++ src/FSAL/FSAL_VFS/os/linux/handle_syscalls.c
@@ -426,7 +426,7 @@ bool vfs_valid_handle(struct gsh_buffdesc *desc)
 	if (!fsid_type_ok) {
 		LogDebug(COMPONENT_FSAL,
 			 "FSID Type %02hhx invalid",
-			 handle0 & HANDLE_FSID_MASK);
+			 (uint8_t) (handle0 & HANDLE_FSID_MASK));
 		return false;
 	}
 
@@ -457,7 +457,7 @@ bool vfs_valid_handle(struct gsh_buffdesc *desc)
 	default:
 		LogDebug(COMPONENT_FSAL,
 			 "Handle Type %02hhx invalid",
-			 handle0 & HANDLE_TYPE_MASK);
+			 (uint8_t) (handle0 & HANDLE_TYPE_MASK));
 		return false;
 	}
